Split api by function and reduce boilerplate
