+++ app/components/rides/services.py
@@ -1,16 +1,9 @@
-from datetime import date, timedelta
-from fastapi import Request
-from fastapi.responses import HTMLResponse
-from humanize import intcomma, metric, ordinal
 from sqlalchemy import ColumnExpressionArgument, false, func, select
 from sqlalchemy.ext.asyncio import AsyncSession
-from sqlalchemy.orm import selectinload
 
-from app.components.core.config import logger, templates
+from app.components.core.config import logger
 from app.components.core.models import Criterion, Ride, User
-from app.components.core.schema import Weather
-from app.components.core.settings import settings
-from app.components.rides.schema import AverageRatings, RideList, RideListItem
+from app.components.rides.schema import AverageRatings
 from app.components.twists.schema import FilterOwnership, FilterWeather, TwistBasic, TwistFilterWithRideOwnership
 
 
@@ -152,167 +145,3 @@ async def calculate_average_rating(
     averages_dict = averages._asdict()  # pyright: ignore [reportPrivateUsage]
 
     return AverageRatings.from_averages(averages_dict, criteria)
-
-
-async def render_averages(
-    request: Request,
-    session: AsyncSession,
-    user: User | None,
-    twist: TwistBasic,
-    filter: TwistFilterWithRideOwnership
-) -> HTMLResponse:
-    """
-    Build and return the TemplateResponse for the ratings averages.
-    """
-    average_ratings = await calculate_average_rating(session, user, twist, filter)
-
-    return templates.TemplateResponse("fragments/rides/averages.html", {
-        "request": request,
-        "twist_id": twist.id,
-        "overall_average": average_ratings.overall,
-        "average_ratings": average_ratings.by_criteria,
-        "criterion_max_value": Criterion.MAX_VALUE
-    })
-
-
-async def render_view_all_button(
-    request: Request,
-    twist_id: int,
-    ride_count: int,
-    filtered: bool
-) -> HTMLResponse:
-    """
-    Build and return the TemplateResponse for the view all rides button.
-    """
-    if ride_count == 0:
-        ride_count_str = "No rides to view"
-    elif ride_count == 1:
-        ride_count_str = "View single ride"
-    elif ride_count > 9999:
-        ride_count_str = "View " + metric(ride_count).replace(" ", "") + " rides"
-    else:
-        ride_count_str = "View " + intcomma(ride_count) + " rides"
-
-    if filtered:
-        ride_count_str += " (filtered)"
-
-    return templates.TemplateResponse("fragments/rides/view_all.html", {
-        "request": request,
-        "twist_id": twist_id,
-        "ride_count": ride_count,
-        "ride_count_str": ride_count_str
-    })
-
-
-async def render_ride_modal(
-    request: Request,
-    twist: TwistBasic,
-    criteria: list[Criterion],
-    today: date
-) -> HTMLResponse:
-    """
-    Build and return the TemplateResponse for the ride modal.
-    """
-    tomorrow = today + timedelta(days=1)
-    criterion_initial_value = int((Criterion.MIN_VALUE + Criterion.MAX_VALUE) / 2)
-
-    return templates.TemplateResponse("fragments/rides/ride_modal.html", {
-        "request": request,
-        "twist": twist,
-        "today": today,
-        "tomorrow": tomorrow,
-        "Weather": Weather,
-        "criteria": criteria,
-        "criterion_min_value": Criterion.MIN_VALUE,
-        "criterion_max_value": Criterion.MAX_VALUE,
-        "criterion_initial_value": criterion_initial_value
-    })
-
-
-async def render_view_modal(
-    request: Request,
-    session: AsyncSession,
-    user: User | None,
-    twist: TwistBasic,
-    filter: TwistFilterWithRideOwnership,
-    offset: int
-) -> HTMLResponse:
-    """
-    Build and return the TemplateResponse for the ride view modal.
-    """
-    # Filtering
-    statement = select(Ride).where(
-        Ride.twist_id == twist.id
-    )
-
-    if filter.ride_ownership == FilterOwnership.OWN:
-        statement = statement.where(Ride.author_id == user.id) if user else statement.where(false())
-
-    weather_conditions = weather_conditions_from(filter.weather)
-    if len(weather_conditions):
-        statement = statement.where(*weather_conditions)
-
-    # Collect rides in order, offset and limited by settings
-    result = await session.scalars(
-        statement
-        .order_by(Ride.date.desc())
-        .offset(offset)
-        .limit(settings.RIDES_FETCHED_PER_QUERY)
-        .options(
-            selectinload(Ride.author).load_only(User.name)
-        )
-    )
-    rides = result.all()
-
-    # Build list items
-    criteria = await Criterion.get_list(session, is_paved=twist.is_paved)
-    ride_list_items: list[RideListItem] = []
-    for ride in rides:
-        # Pre-format the date for easier display in the template
-        ordinal_day = ordinal(ride.date.day)
-        formatted_date = ride.date.strftime(f"%B {ordinal_day}, %Y")
-
-        # Set author name whether they exist or not
-        author_name = ride.author.name if ride.author else settings.DELETED_USER_NAME
-
-        # Check if the user is allowed to delete the ride
-        editable = (user.is_superuser or user.id == ride.author_id) if user else False
-
-        ride_list_items.append(RideListItem(
-            id=ride.id,
-            author_name=author_name,
-            editable=editable,
-            formatted_date=formatted_date,
-            weather=ride.weather,
-            ratings={
-                c.slug: ride.ratings[c.slug]
-                for c in criteria
-            }
-        ))
-
-    # Compile list with criteria descriptions
-    ride_list = RideList(
-        criteria_descriptions={
-            c.slug: c.description or ""
-            for c in criteria
-        },
-        items=ride_list_items
-    )
-
-    # For offset 0 (beginning), use the template including the header, otherwise just the list
-    if offset == 0:
-        return templates.TemplateResponse("fragments/rides/view_modal.html", {
-            "request": request,
-            "twist": twist,
-            "ride_list": ride_list,
-            "next_offset": offset + settings.RIDES_FETCHED_PER_QUERY,
-            "criterion_max_value": Criterion.MAX_VALUE
-        })
-    else:
-        return templates.TemplateResponse("fragments/rides/view_list.html", {
-            "request": request,
-            "twist": twist,
-            "ride_list": ride_list,
-            "next_offset": offset + settings.RIDES_FETCHED_PER_QUERY,
-            "criterion_max_value": Criterion.MAX_VALUE
-        })
