+++ app/main.py
@@ -12,9 +12,7 @@ from time import time
 from typing import Awaitable, Callable, cast
 import uvicorn
 
-from app.components.admin import api as admin
 from app.components.admin.services import create_first_admin
-from app.components.auth import api as auth
 from app.components.auth.services import login_and_set_response_cookie
 from app.components.core.config import logger, tags_metadata, templates
 from app.components.core.database import apply_migrations, create_automigration, get_db, wait_for_db
@@ -23,12 +21,15 @@ from app.components.core.models import User
 from app.components.core.redis_client import redis_client
 from app.components.core.settings import Settings, settings
 from app.components.core.utility import format_loc_for_user, raise_http, sort_schema_names
-from app.components.debug import api as debug
-from app.components.rides import api as rides
 from app.components.rides.services import initialize_criteria
-from app.components.twists import api as twists
-from app.components.users import api as users
 from app.components.users.services import current_user_optional
+
+from app.components.admin import api as admin_api, fragments as admin_fragments
+from app.components.auth import api as auth_api, fragments as auth_fragments
+from app.components.debug import api as debug_api, fragments as debug_fragments
+from app.components.rides import api as rides_api, fragments as rides_fragments
+from app.components.twists import api as twists_api, fragments as twists_fragments
+from app.components.users import api as users_api, fragments as users_fragments
 
 
 @asynccontextmanager
@@ -221,17 +222,24 @@ async def get_latest_version(request: Request) -> HTMLResponse:
     return HTMLResponse(content=f"<strong>{latest_version}</strong>")
 
 
-app.include_router(admin.router)
-app.include_router(auth.router)
-app.include_router(debug.router)
-app.include_router(rides.router)
-app.include_router(twists.router)
-app.include_router(users.router)
+app.include_router(admin_api.router)
+app.include_router(admin_fragments.router)
+app.include_router(auth_api.router)
+app.include_router(auth_fragments.router)
+app.include_router(debug_api.router)
+app.include_router(debug_fragments.router)
+app.include_router(rides_api.router)
+app.include_router(rides_fragments.router)
+app.include_router(twists_api.router)
+app.include_router(twists_fragments.router)
+app.include_router(users_api.router)
+app.include_router(users_fragments.router)
 
 # TODO: fix update_schema_name no longer working
 # update_schema_name(app, auth.login, "UserLoginForm")
 # update_schema_name(app, debug.load_state, "StateLoadUploadFile")
 sort_schema_names(app)
+
 
 if __name__ == "__main__":
     wait_for_db()
