+++ app/templates/debug.html
@@ -43,7 +43,7 @@
                 <button type="submit" class="button button-block">Load State</button>
             </div>
         </form>
-        <form class="modal-form" method="POST" action="/debug/save">
+        <form class="modal-form" method="GET" action="/debug/save">
             <div class="button-group">
                 <span class="button-group-label">Save the database state to a JSON file</span>
                 <button type="submit" class="button button-block">Save State</button>
